infrastructure/expected-fail/expected-intermittent.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Intermittent test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(Math.random() > 0.5)
});
</script>
